% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket_problem.R
\name{bucket_problem}
\alias{bucket_problem}
\title{Create a bucket problem (experimental).}
\usage{
bucket_problem(initial, text = c("Drag from here",
  "Construct your solution here"), header = NULL, input_id, group_name,
  class = "default-sortable default-bucket",
  options = sortable_options(), orientation = c("horizontal",
  "vertical"))
}
\arguments{
\item{initial}{Vector with initial values for problem (to appear in left
column).  Note: this must be a super-set of all answers.}

\item{text}{Vector of headings for each column.}

\item{header}{Text that appears at the top of the bucket list.  (This is
encoded as an HTML \code{<p>} tag, so not strictly speaking a header.)}

\item{input_id}{Character vector of \code{input_id} to pass (individually) to
\code{\link[=rank_list]{rank_list()}}.}

\item{group_name}{Passed to \code{SortableJS} as the group name. Also the input
value set in Shiny. (\code{input[[group_name]]})}

\item{class}{A css class applied to the rank list.  This can be used to
define custom styling.}

\item{options}{Options to be supplied to \link{sortable_js} object. See \link{sortable_options} for more details}

\item{orientation}{Either \code{horizontal} or \code{vertical}, and specifies the
layout of the components on the page.}
}
\description{
This function implements the bucket problem, as exposed by
\code{\link[=question_bucket]{question_bucket()}}. Most users will only use this function inside a
\code{learnr} tutorial, so please see the documentation at \code{\link[=question_bucket]{question_bucket()}}
}
