% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css.R
\name{css_parsons}
\alias{css_parsons}
\title{CSS stylesheets.}
\usage{
css_parsons()
}
\value{
Each of these functions return a character string with valid CSS.
}
\description{
These functions provide cascading style sheets (CSS) for:
}
\details{
\itemize{
\item \code{css_parsons()} : \code{\link[=parsons]{parsons()}}
}
}
\examples{
cat(
  css_parsons()
)

}
