% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsons_expectations.R
\name{expectation}
\alias{expectation}
\title{Add expectations to a parsons problem.}
\usage{
expectation(fun, msg = "failure")
}
\arguments{
\item{fun}{A function of x that should evaluate to TRUE or FALSE}

\item{msg}{Message to display if \code{fun} evaluates to TRUE}
}
\description{
Add expectations to a parsons problem.
}
