% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectations.R
\name{expectation_pass}
\alias{expectation_pass}
\title{Add expectations to a parsons problem.}
\usage{
expectation_pass(fun, message = "failure")
}
\arguments{
\item{fun}{A function of x that should evaluate to TRUE or FALSE}

\item{message}{Message to display if \code{fun} evaluates to TRUE}
}
\description{
Add expectations to a parsons problem.
}
