% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectations.R
\name{pass_if}
\alias{pass_if}
\alias{fail_if}
\alias{message_if}
\alias{contains_all}
\alias{contains_any}
\title{Add expectations to a parsons problem.}
\usage{
pass_if(f, message = NULL)

fail_if(f, message = "Incorrect")

message_if(f)

contains_all(x, ...)

contains_any(x, ...)
}
\arguments{
\item{f}{One of:
\itemize{
\item A character vector, indicating an exact match
\item A function of the function \code{function(x){...}} that evaluates to TRUE or FALSE
\item A function of the form \code{~ .}, as used by the tidy evaluation, e.g. in \link[purrr:map]{purrr::map}
}}

\item{message}{Message to display if \code{fun} evaluates to TRUE}

\item{x}{Answer to test}

\item{...}{Combined into an answer list}
}
\description{
Add expectations to a parsons problem.
}
