% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsons-package.R
\docType{package}
\name{parsons}
\alias{parsons}
\alias{_PACKAGE}
\alias{parsons-package}
\title{parsons: Create Parsons Programming Problems in 'learnr` Tutorials}
\description{
Create Parsons Programming Problems in 'learnr' Tutorials.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstudio/parsons}
  \item Report bugs at \url{https://github.com/rstudio/parsons/issues}
}

}
\author{
\strong{Maintainer}: Andrie de Vries \email{apdevries@gmail.com}

Other contributors:
\itemize{
  \item Barret Schloerke \email{barret@rstudio.com} [contributor]
}

}
