% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_bucket.R
\name{question_bucket}
\alias{question_bucket}
\title{bucket problem question for learnr tutorials (experimental).}
\usage{
question_bucket(initial, ..., text = c("Drag from here",
  "Construct your solution here"), orientation = c("horizontal",
  "vertical"), correct = "Correct!", incorrect = "Incorrect",
  try_again = incorrect, message = NULL, post_message = NULL,
  loading = c("Loading: "), submit_button = "Give feedback",
  try_again_button = "Try Again", allow_retry = TRUE,
  random_answer_order = TRUE, options = sortable_options())
}
\arguments{
\item{initial}{Initial value of answer options. This must be a character vector.}

\item{...}{One or more answers.  Passed to \code{\link[learnr:question]{learnr::question()}}.}

\item{text}{Question or option text}

\item{orientation}{Either \code{horizontal} or \code{vertical}, and specifies the
layout of the components on the page.}

\item{correct}{Text to print for a correct answer (defaults to "Correct!")}

\item{incorrect}{Text to print for an incorrect answer (defaults to "Incorrect")
when \code{allow_retry} is \code{FALSE}.}

\item{try_again}{Text to print for an incorrect answer (defaults to "Incorrect")
when \code{allow_retry} is \code{TRUE}.}

\item{message}{Additional message to display along with correct/incorrect feedback.
This message is always displayed after a question submission.}

\item{post_message}{Additional message to display along with correct/incorrect feedback.
If \code{allow_retry} is \code{TRUE}, this message will only be displayed after the
correct submission.  If \code{allow_retry} is \code{FALSE}, it will produce a second
message alongside the \code{message} message value.}

\item{loading}{Loading text to display as a placeholder while the question is loaded}

\item{submit_button}{Label for the submit button. Defaults to \code{"Submit Answer"}}

\item{try_again_button}{Label for the try again button. Defaults to \code{"Submit Answer"}}

\item{allow_retry}{Allow retry for incorrect answers. Defaults to \code{FALSE}.}

\item{random_answer_order}{Display answers in a random order.}

\item{options}{Extra options to be stored in the question object.}
}
\description{
bucket problem question for learnr tutorials (experimental).
}
\examples{
## Example of bucket problem inside a learn tutorial
if (interactive()) {
  learnr::run_tutorial("bucket", package = "parsons")
}
}
