% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_parsons.R
\name{question_try_again_ui.parsons_question}
\alias{question_try_again_ui.parsons_question}
\title{Disable input after student submitted answer.}
\usage{
\method{question_try_again_ui}{parsons_question}(question, answer_input,
  ...)
}
\arguments{
\item{question}{Question object}

\item{answer_input}{user input value}

\item{...}{not used}
}
\description{
Disable input after student submitted answer.
}
