% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_bucket.R
\name{question_ui_try_again.bucket_question}
\alias{question_ui_try_again.bucket_question}
\title{Disable input after student submitted answer.}
\usage{
\method{question_ui_try_again}{bucket_question}(question, answer_input,
  ...)
}
\arguments{
\item{question}{Question object}

\item{answer_input}{user input value}

\item{...}{not used}
}
\description{
Disable input after student submitted answer.
}
