% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsons_expectations.R
\name{pass_if}
\alias{pass_if}
\alias{fail_if}
\title{Add expectations to a parsons problem.}
\usage{
pass_if(f, message = NULL)

fail_if(f, message = "Incorrect")
}
\arguments{
\item{f}{One of:
\itemize{
\item A character vector, indicating an exact match
\item A function of the function \code{function(x){...}} that evaluates to TRUE or FALSE
\item A function of the form \code{~ .}, as used by the tidy evaluation, e.g. in \link[purrr:map]{purrr::map}
}}

\item{message}{Message to display if \code{fun} evaluates to TRUE}
}
\description{
Add expectations to a parsons problem.
}
